// src/app/shared/model/modal-data.model.ts
/**
 * © Rolling Array https://rollingarray.co.in/
 *
 *
 * @summary Modal data
 * @author code@rollingarray.co.in
 *
 * Created at     : 2021-05-18 19:11:57 
 * Last modified  : 2021-05-18 19:11:57 
 */


export interface ModalData{
  cancelled: boolean;
  operationSubmitted: boolean;
  returnMessage ?: string;
}