// src/app/shared/enum/operations.enum.ts
/**
 * © Rolling Array https://rollingarray.co.in/
 *
 *
 * @summary Operations enum
 * @author code@rollingarray.co.in
 *
 * Created at     : 2021-04-29 11:22:01 
 * Last modified  : 2021-08-10 19:34:48
 */


export enum OperationsEnum {
	Create = "create",
	Edit = "edit",
	Delete = "delete",
	Start = 'start',
	Stop = 'stop',
	Add = 'add',
	Remove = 'remove',
	Approve = "APPROVE",
	Reject = "REJECT",
	Pending = "PENDING",
	Accept = "ACCEPT",
	Received = "RECEIVED",
	Inactive = "INACTIVE",
	Active = "Active",
	Resume = "Resume",
	Closed = "Closed",
	Future = "Future",
}